requirements.txt
tensorflow==2.15.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
numpy==1.24.3
pillow==10.2.0
python-jose==3.3.0
python-dotenv==1.0.0
scikit-learn==1.3.0
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.3
pytest==7.4.0
requests==2.31.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 